% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SAMGEP-package}
\alias{SAMGEP-package}
\title{SAMGEP: A Semi-supervised Method for Prediction of Phenotype Event Times Using the Electronic Health Record}
\description{
Semi-supervised Adaptive Markov Gaussian Embedding Process (SAMGEP) is a novel semi-supervised machine learning algorithm to predict phenotype event times using Electronic Health Record (EHR) data.
}
\keyword{package}
