% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulate}
\alias{simulate}
\title{Simulated Dataset}
\format{
An object of class \code{function} of length 1.
}
\usage{
simulate
}
\description{
Click \href{https://github.com/celehs/SAMGEP/blob/master/data-raw/simdata.R}{HERE} to view details.
}
\examples{
str(simulate)
}
\keyword{datasets}
